layouts/partials: head: Add nofollow when adding noindex.
It is the natural intuition for a user that crawlers like Google won't
index pages linked by pages which are marked as noindex.

But it turns out crawlers are snarky and will follow the links in a
noindex page. We need to set nofollow too to stop that from happening.

Signed-off-by: Siddh Raman Pant <25429745+siddhpant@users.noreply.github.com>